+++ magma/lib/magma/query/subselect_count.rb
@@ -20,7 +20,7 @@ class Magma
         restrict_constraints
       )
 
-      apply_filter_constraints(query)
+      apply_filters(query)
     end
 
     def count_column
